requirements.txt
streamlit
yfinance
requests
beautifulsoup4
lxml
transformers
torch
scikit-learn
plotly
googlesearch-python
